+++ src/days/day03.rs
@@ -25,7 +25,7 @@ mod tests {
     }
 }
 
-pub fn solve() -> (i32,i32) {
+pub fn solve() -> (i64,i64) {
     let result = do_puzzle("day3_1.txt");
 
     match result{
@@ -36,7 +36,7 @@ pub fn solve() -> (i32,i32) {
     (0,0)
 }
 
-fn do_puzzle(input: &str)-> Result<(i32,i32), io::Error>{
+fn do_puzzle(input: &str)-> Result<(i64,i64), io::Error>{
     let contents = utils::read_file(input)?;
     let extra_line = '.'.to_string().repeat(10);
 
@@ -56,7 +56,7 @@ fn do_puzzle(input: &str)-> Result<(i32,i32), io::Error>{
     Ok(result)
 }
 
-fn find_num_in_line(input: &Vec<Vec<char>>) -> (i32,i32) {
+fn find_num_in_line(input: &Vec<Vec<char>>) -> (i64,i64) {
     let mut result = 0;
     let mut numbers: Vec<(i32, bool, Vec<i32>)> = Vec::new();
     let mut gears_vals: HashMap<i32,Vec<i32>> = HashMap::new();
@@ -123,5 +123,5 @@ fn find_num_in_line(input: &Vec<Vec<char>>) -> (i32,i32) {
             }
         }
     }
-    (result,result2)
+    (result as i64,result2 as i64)
 }
