+++ src/days/day04.rs
@@ -41,7 +41,7 @@ mod tests {
     }
 }
 
-pub fn solve() -> (i32,i32) {
+pub fn solve() -> (i64,i64) {
     let result = do_puzzle("day4_1.txt");
     match result{
         Ok(value) => {return value}
@@ -51,7 +51,7 @@ pub fn solve() -> (i32,i32) {
     (0,0)
 }
 
-fn do_puzzle(input: &str)-> Result<(i32,i32), io::Error>{
+fn do_puzzle(input: &str)-> Result<(i64,i64), io::Error>{
     let contents = utils::read_file(input)?;
     let mut num_cards: Vec<i32> = vec![1; contents.lines().count()];
     let mut result1 = 0;
@@ -72,7 +72,7 @@ fn do_puzzle(input: &str)-> Result<(i32,i32), io::Error>{
         
         i += 1;
     }
-    Ok((result1,result2))
+    Ok((result1 as i64,result2 as i64))
 }
 
 fn find_nums(input: &str) -> (i32,i32) {
